8321 extract payment alert view model creation
